add Pagination.new/2
