+++ lib/flop_phoenix/pagination.ex
@@ -4,6 +4,7 @@ defmodule Flop.Phoenix.Pagination do
   component.
   """
 
+  alias Flop.Meta
   alias Flop.Phoenix.Misc
 
   require Logger
@@ -27,7 +28,6 @@ defmodule Flop.Phoenix.Pagination do
     to that page that also includes query parameters for filters and sorting.
   - `previous_page`
   - `total_pages`
-
 
   ## For cursor-based pagination
 
@@ -80,6 +80,180 @@ defmodule Flop.Phoenix.Pagination do
     next_direction: :next,
     previous_direction: :previous
   ]
+
+  @doc """
+  Returns a `Pagination` struct for the given `Flop.Meta` struct.
+  """
+  @spec new(Meta.t(), keyword) :: t
+  def new(meta, opts \\ [])
+
+  def new(
+        %Flop.Meta{
+          errors: [],
+          has_next_page?: has_next_page?,
+          has_previous_page?: has_previous_page?,
+          total_pages: total_pages
+        } = meta,
+        opts
+      )
+      when has_next_page? or has_previous_page? do
+    opts = Keyword.validate!(opts, page_links: 5, path: nil, reverse: false)
+    page_links = Keyword.fetch!(opts, :page_links)
+    path = Keyword.fetch!(opts, :path)
+    reverse = Keyword.fetch!(opts, :reverse)
+    pagination_type = pagination_type(meta.flop)
+
+    if pagination_type in [:page, :offset] do
+      path_fun = build_page_path_fun(meta, path)
+
+      {page_range_start, page_range_end} =
+        Flop.Phoenix.page_link_range(
+          page_links,
+          meta.current_page,
+          meta.total_pages
+        )
+
+      %__MODULE__{
+        current_page: meta.current_page,
+        ellipsis_end?: page_range_end < meta.total_pages - 1,
+        ellipsis_start?: page_range_start > 2,
+        next_page: meta.next_page,
+        page_range_end: page_range_end,
+        path_fun: path_fun,
+        page_range_start: page_range_start,
+        pagination_type: pagination_type,
+        previous_page: meta.previous_page,
+        total_pages: total_pages
+      }
+    else
+      previous_cursor = if has_previous_page?, do: meta.start_cursor
+      next_cursor = if has_next_page?, do: meta.end_cursor
+      path_fun = build_cursor_path_fun(meta, path)
+
+      {previous_direction, previous_cursor, next_direction, next_cursor} =
+        if reverse do
+          {:next, next_cursor, :previous, previous_cursor}
+        else
+          {:previous, previous_cursor, :next, next_cursor}
+        end
+
+      %__MODULE__{
+        next_cursor: next_cursor,
+        next_direction: next_direction,
+        path_fun: path_fun,
+        pagination_type: pagination_type,
+        previous_cursor: previous_cursor,
+        previous_direction: previous_direction
+      }
+    end
+  end
+
+  def new(%Meta{flop: flop}, _) do
+    %__MODULE__{pagination_type: pagination_type(flop)}
+  end
+
+  defp pagination_type(%Flop{limit: limit, offset: offset})
+       when is_integer(limit) and is_integer(offset) do
+    :offset
+  end
+
+  defp pagination_type(%Flop{page: page, page_size: page_size})
+       when is_integer(page) and is_integer(page_size) do
+    :page
+  end
+
+  defp pagination_type(%Flop{first: first}) when is_integer(first) do
+    :first
+  end
+
+  defp pagination_type(%Flop{last: last}) when is_integer(last) do
+    :last
+  end
+
+  defp build_page_path_fun(_meta, nil), do: fn _ -> nil end
+
+  defp build_page_path_fun(meta, path) do
+    &build_page_path(path, &1, build_page_query_params(meta))
+  end
+
+  defp build_page_query_params(meta) do
+    meta.flop
+    |> ensure_page_based_params()
+    |> Flop.Phoenix.to_query(backend: meta.backend, for: meta.schema)
+  end
+
+  defp build_page_path(path, page, query_params) do
+    Flop.Phoenix.build_path(path, maybe_put_page(query_params, page))
+  end
+
+  defp ensure_page_based_params(%Flop{} = flop) do
+    %{
+      flop
+      | after: nil,
+        before: nil,
+        first: nil,
+        last: nil,
+        limit: nil,
+        offset: nil,
+        page_size: flop.page_size || flop.limit,
+        page: flop.page
+    }
+  end
+
+  defp maybe_put_page(params, 1), do: Keyword.delete(params, :page)
+  defp maybe_put_page(params, page), do: Keyword.put(params, :page, page)
+
+  defp build_cursor_path_fun(_meta, nil), do: fn _, _ -> nil end
+
+  defp build_cursor_path_fun(meta, path) do
+    &build_cursor_path(path, &1, &2, build_cursor_query_params(meta))
+  end
+
+  defp build_cursor_query_params(meta) do
+    meta.flop
+    |> ensure_cursor_based_params()
+    |> Flop.Phoenix.to_query(backend: meta.backend, for: meta.schema)
+  end
+
+  defp build_cursor_path(path, cursor, direction, query_params) do
+    Flop.Phoenix.build_path(
+      path,
+      maybe_put_cursor(query_params, cursor, direction)
+    )
+  end
+
+  defp ensure_cursor_based_params(%Flop{} = flop) do
+    %{
+      flop
+      | after: nil,
+        before: nil,
+        limit: nil,
+        offset: nil,
+        page_size: nil,
+        page: nil
+    }
+  end
+
+  defp maybe_put_cursor(query_params, nil, _), do: query_params
+
+  defp maybe_put_cursor(query_params, cursor, :previous) do
+    query_params
+    |> Keyword.merge(
+      before: cursor,
+      last: query_params[:last] || query_params[:first],
+      first: nil
+    )
+    |> Keyword.delete(:first)
+  end
+
+  defp maybe_put_cursor(query_params, cursor, :next) do
+    query_params
+    |> Keyword.merge(
+      after: cursor,
+      first: query_params[:first] || query_params[:last]
+    )
+    |> Keyword.delete(:last)
+  end
 
   @doc false
   @spec default_opts() :: [Flop.Phoenix.pagination_option()]
